+++ app/src/main/kotlin/com/example/killersudoku/data/di/DataModule.kt
@@ -35,6 +35,7 @@ object DataModule {
         )
             .addMigrations(AppDatabase.MIGRATION_1_2)
             .addMigrations(AppDatabase.MIGRATION_2_3)
+            .addMigrations(AppDatabase.MIGRATION_3_4)
             .build()
 
     @Provides